+++ src/main/java/com/uvasoftware/scanii/impl/ErrorExtractor.java
@@ -4,6 +4,10 @@ import com.uvasoftware.scanii.internal.JSON;
 
 class ErrorExtractor {
   String extract(String contents) {
-    return JSON.load(contents).get("error").asText();
+    var js =  JSON.load(contents);
+    if (js.has("error")) {
+      return js.get("error").asText();
+    }
+    return contents;
   }
 }